Replace BigDecimal price with long
Теперь вроде все однообразно...

+++ order-service/src/main/java/orbmrkt/order/dto/CreateOrderRequest.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import orbmrkt.dto.ProductType;
 
-import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
@@ -13,7 +12,7 @@ public class CreateOrderRequest {
     @JsonProperty("product_type")
     private ProductType productType;
 
-    private BigDecimal price;
+    private long price;
 
     private Map<String, Object> payload;
 }
